IOS Layout finalize/fonts fix
