+++ packages/client-core/src/socialmedia/components/popups/FeedPopup/FeedPopup.module.scss
@@ -1,4 +1,4 @@
-.isIOS{
+.isIos{
 	background-color: rgb(255,255,255);
 	.feedPageIosWrapper{
 		height: calc(100vh - 3rem);
