+++ packages/social/src/pages/index.module.scss
@@ -1,14 +1,14 @@
 @font-face{
     font-family: "SFProText-Semibold";
-    src : url("assets/fonts/FontsFree-Net-SFProText-Semibold-1.ttf");
+    src : url("/assets/fonts/FontsFree-Net-SFProText-Semibold-1.ttf");
   }
   @font-face{
     font-family: "SFProText-Regular";
-    src : url("assets/fonts/FontsFree-Net-SFProText-Regular-1.ttf");
+    src : url("/assets/fonts/FontsFree-Net-SFProText-Regular-1.ttf");
   }
   @font-face{
     font-family: "SFProText-Heavy";
-    src : url("assets/fonts/FontsFree-Net-SFProText-Heavy.ttf");
+    src : url("/assets/fonts/FontsFree-Net-SFProText-Heavy.ttf");
   }
 body{
   font-size: 13px;
