+++ packages/social/src/pages/index.tsx
@@ -31,6 +31,7 @@ import styles from './index.module.scss';
 import image from '/static/images/image.jpg';
 import mockupIPhone from '/static/images/mockupIPhone.jpg';
 import Splash from '@xrengine/client-core/src/socialmedia/components/Splash';
+import {isIOS} from "@xrengine/client-core/src/util/platformCheck";
 
 const mapStateToProps = (state: any): any => {
   return {
@@ -84,6 +85,7 @@ const  Home = ({ createCreator,  doLoginAuto, auth, creatorsState, webxrnativeSt
     setFeedOnborded(false);
     setFeedHintsOnborded(false);
   };
+  const platformClass = isIOS ? styles.isIos : "";
 
   if(!currentCreator || currentCreator === null) return <Splash />;
 
@@ -91,10 +93,10 @@ const  Home = ({ createCreator,  doLoginAuto, auth, creatorsState, webxrnativeSt
 
   if(!onborded) return <Onboard setOnborded={changeOnboarding} image={image} mockupIPhone={mockupIPhone} />;
 
-  const isIOS = /iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream;
 
 
-  return (<div className={isIOS ? styles.isIos : ''}>
+
+  return (<div className={platformClass}>
     {!feedOnborded && <FeedOnboarding setFeedOnborded={setFeedOnborded} />}
     <div className={webxrRecorderActivity ? styles.hideContent+' '+styles.viewport : styles.viewport}>
         <AppHeader logo="/assets/logoBlack.png" />
